fim da aula 3

+++ server/src/routes/memories.ts
@@ -3,8 +3,19 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-	app.get("/memories", async () => {
+	app.addHook("preHandler", async (request) => {
+		await request.jwtVerify();
+	});
+
+	app.get("/memories", async (request) => {
+		await request.jwtVerify();
+
+		const {} = request.user;
+
 		const memories = await prisma.memory.findMany({
+			where: {
+				userId: request.user.sub,
+			},
 			orderBy: {
 				createdAt: "asc",
 			},
@@ -19,7 +30,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		});
 	});
 
-	app.get("/memories/:id", async (request) => {
+	app.get("/memories/:id", async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -31,6 +42,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
 				id,
 			},
 		});
+
+		if (!memory.isPublic && memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
+
 		return memory;
 	});
 
@@ -48,29 +64,39 @@ export async function memoriesRoutes(app: FastifyInstance) {
 				content,
 				coverUrl,
 				isPublic,
-				userId: "a46e1d73-4cd6-4000-b730-37c26af23759",
+				userId: request.user.sub,
 			},
 		});
 
 		return memory;
 	});
 
-	app.put("/memories/:id", async (request) => {
+	app.put("/memories/:id", async (request, reply) => {
+		const paramsSchema = z.object({
+			id: z.string().uuid(),
+		});
+
+		const { id } = paramsSchema.parse(request.params);
+
 		const bodySchema = z.object({
 			content: z.string(),
 			coverUrl: z.string(),
 			isPublic: z.coerce.boolean().default(false),
 		});
 
-		const paramsSchema = z.object({
-			id: z.string().uuid(),
-		});
-
 		const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-		const { id } = paramsSchema.parse(request.params);
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
 
-		const memory = await prisma.memory.update({
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
+
+		memory = await prisma.memory.update({
 			where: {
 				id,
 			},
@@ -84,14 +110,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		return memory;
 	});
 
-	app.delete("/memories/:id", async (request) => {
+	app.delete("/memories/:id", async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
 
 		const { id } = paramsSchema.parse(request.params);
 
-		const memory = await prisma.memory.delete({
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
+
+		await prisma.memory.delete({
 			where: {
 				id,
 			},
